Create usweather.sh
For this command to actually display the weather, please note that this command is specifically designed for the United States. All calculations will be in the imperial system with this default implementation that includes a log of certain past weather alerts.

If your location is not available, you may not be in the United States or you may not have typed the specific type of place ("city", "town", "CCD", etc.).